+++ applications/launchpad/backend/src/docker/error.rs
@@ -48,6 +48,7 @@ pub enum DockerWrapperError {
 }
 
 impl DockerWrapperError {
+    /// Combine all error messages down the chain into one string.
     pub fn chained_message(&self) -> String {
         let mut messages = vec![self.to_string()];
         let mut this = self as &dyn Error;
