+++ applications/launchpad/backend/src/docker/mod.rs
@@ -32,18 +32,16 @@ pub mod helpers;
 
 pub use error::DockerWrapperError;
 pub use filesystem::create_workspace_folders;
-pub use models::{ContainerId, ContainerState, ContainerStatus, LogMessage};
+pub use models::{ContainerId, ContainerState, ContainerStatus,    ImageType, LogMessage,  TariNetwork};
 pub use settings::{
     BaseNodeConfig,
-    ImageType,
     LaunchpadConfig,
     MmProxyConfig,
     Sha3MinerConfig,
-    TariNetwork,
     WalletConfig,
     XmRigConfig,
     DEFAULT_MINING_ADDRESS,
     DEFAULT_MONEROD_URL
 };
-pub use workspace::{TariWorkspace, Workspaces, DEFAULT_IMAGES};
+pub use workspace::{TariWorkspace, Workspaces};
 pub use wrapper::DockerWrapper;
